Code separation
- Use of controllers
- Split Game view code into several components
