+++ imports/startup/client/controllers/GamesListController.js
@@ -1,0 +1,13 @@
+import {Meteor} from 'meteor/meteor';
+import {Games} from '/imports/api/games/games.js';
+
+export const GamesListController = RouteController.extend({
+	waitOn: function() {
+		return Meteor.subscribe('games');
+	},
+	data: function() {
+		return {
+			games: Games.find({}, {sort: [['createdAt', 'desc']]})
+		};
+	}
+});
