+++ imports/ui/components/afterGame.html
@@ -1,0 +1,53 @@
+<template name="afterGame">
+	<div id="afterGameContainer" class="after-game-container">
+		<div>
+			<div>
+				<div class="after-game-item after-game-title">{{getAfterGameTitle}}</div>
+
+				{{#if showEloChangeAfterGame}}
+					<div class="after-game-item after-game-elo">
+						{{#each players}}
+							<div>
+								<div class="player-name">{{name}}</div>
+								<div>{{getPlayerEloRating this ../eloScores}}</div>
+								<div>{{> eloRatingChange eloRatingChange=(getPlayerGameEloRatingChange this ../eloScores)}}</div>
+							</div>
+						{{/each}}
+					</div>
+				{{/if}}
+
+				{{#if isGamePlayer}}
+					<div class="after-game-item after-game-actions">
+						{{#if showOpponentDeclinedRematch}}
+							<div class="rematch-declined">Your opponent declined a rematch...</div>
+						{{/if}}
+						{{#if showPlayerLeftTheGame}}
+							<div class="rematch-declined">Your opponent left the game...</div>
+						{{/if}}
+						{{#if showWaitingForReply}}
+							<div>Waiting for your opponent to accept...</div>
+						{{/if}}
+						{{#if showCreatingGame}}
+							<div>Creating game...</div>
+						{{/if}}
+						{{#if askForRematchReply}}
+							<div>
+								Your opponent is asking for a rematch. Do you accept?<br />
+								<a href="#" class="button" data-action="game-rematch">Yes</a>
+								<a href="#" class="button" data-action="declined-game-rematch">No</a>
+							</div>
+						{{/if}}
+						{{#if showActionAfterGame}}
+							<strong>What do you want to do next?</strong>
+							{{#if showAskForRematch}}
+								<div><a href="#" data-action="game-rematch">Ask for a rematch</a></div>
+							{{/if}}
+							<div><a href="{{pathFor 'home'}}">Go to the homepage</a></div>
+							<div><a href="{{pathFor 'rank'}}">View rank</a></div>
+						{{/if}}
+					</div>
+				{{/if}}
+			</div>
+		</div>
+	</div>
+</template>
