+++ imports/ui/components/afterGame.js
@@ -1,0 +1,141 @@
+import {Meteor} from 'meteor/meteor';
+import {Template} from 'meteor/templating';
+import {Session} from 'meteor/session';
+import {Players} from '/imports/api/games/players.js';
+import {
+	isGamePlayer,
+	isGameStatusFinished,
+	isGameStatusTimeout,
+	getWinnerName
+} from '/imports/api/games/utils.js';
+import {
+	playerAcceptedRematch,
+	playerDeclinedRematch,
+	playerHasNotRepliedRematch,
+	playerLeftGame,
+	currentPlayerHasRepliedRematch,
+	currentPlayerAcceptedRematch
+} from '/imports/lib/client/gameSetup.js';
+
+import './afterGame.html';
+
+Template.afterGame.helpers({
+	isGamePlayer: function() {
+		return isGamePlayer(Session.get('game'));
+	},
+
+	showEloChangeAfterGame: function() {
+		return !this.game.isPracticeGame && isGameStatusFinished(this.game.status);
+	},
+
+	getAfterGameTitle: function() {
+		if (isGameStatusTimeout(this.game.status)) {
+			return 'The game has timed out...';
+		} else if (isGameStatusFinished(this.game.status)) {
+			return getWinnerName(this.game) + ' wins';
+		}
+	},
+
+	getPlayerEloRating: function(player, eloScores) {
+		let eloRating = '-';
+
+		eloScores.forEach((eloScore) => {
+			if (player.userId === eloScore.userId) {
+				eloRating = eloScore.eloRating;
+			}
+		});
+
+		return eloRating;
+	},
+
+	getPlayerGameEloRatingChange: function(player, eloScores) {
+		let gameEloRatingChange = null;
+
+		eloScores.forEach((eloScore) => {
+			if (player.userId === eloScore.userId) {
+				gameEloRatingChange = eloScore.eloRatingChange;
+			}
+		});
+
+		return gameEloRatingChange;
+	},
+
+	showActionAfterGame() {
+		const players = Players.find({gameId: Session.get('game')});
+
+		return (
+			!playerAcceptedRematch(players) ||
+			playerDeclinedRematch(players) ||
+			playerLeftGame(players)
+		);
+	},
+
+	showAskForRematch() {
+		const players = Players.find({gameId: Session.get('game')});
+
+		return (
+			!playerAcceptedRematch(players) &&
+			!playerDeclinedRematch(players) &&
+			!playerLeftGame(players)
+		);
+	},
+
+	showWaitingForReply() {
+		const players = Players.find({gameId: Session.get('game')});
+
+		return (
+			currentPlayerAcceptedRematch(players, Meteor.userId()) &&
+			playerHasNotRepliedRematch(players) &&
+			!playerLeftGame(players)
+		);
+	},
+
+	showPlayerLeftTheGame() {
+		const players = Players.find({gameId: Session.get('game')});
+
+		return (
+			playerLeftGame(players) &&
+			!playerDeclinedRematch(players)
+		);
+	},
+
+	showOpponentDeclinedRematch() {
+		const players = Players.find({gameId: Session.get('game')});
+
+		return (
+			currentPlayerAcceptedRematch(players, Meteor.userId()) &&
+			playerDeclinedRematch(players)
+		);
+	},
+
+	askForRematchReply() {
+		const players = Players.find({gameId: Session.get('game')});
+
+		return (
+			playerAcceptedRematch(players) &&
+			!currentPlayerHasRepliedRematch(players, Meteor.userId()) &&
+			!playerDeclinedRematch(players) &&
+			!playerLeftGame(players)
+		);
+	},
+
+	showCreatingGame() {
+		const players = Players.find({gameId: Session.get('game')});
+
+		return (
+			!playerDeclinedRematch(players) &&
+			!playerHasNotRepliedRematch(players)
+		);
+	}
+});
+
+Template.afterGame.events({
+	'click [data-action="game-rematch"]': function() {
+		Meteor.call('replyRematch', Session.get('game'), true);
+	},
+
+	'click [data-action="declined-game-rematch"]': function() {
+		Meteor.call('replyRematch', Session.get('game'), false);
+	}
+});
+
